Corrige el calculo del consecutivo con prefijo y aisla los tests del worktree
SUBSTRING(id FROM $1) con el parametro sin castear hacia que Postgres
resolviera la variante de REGEX en vez de la de posicion: extraia el digito
encontrado dentro del ID en lugar del sufijo, de modo que el generador
devolvia PER005 con 328 perifericos ya cargados. Se castea a ::int en el
repositorio y en el script de migracion. El script no se vio afectado en su
corrida real porque no habia ningun PER previo y arrancaba desde cero.

Se agrega IncrementalId.test con los casos que lo habrian detectado:
secuencia, hueco en la numeracion, salto de tres a cuatro digitos y
convivencia con categorias sin prefijo.

jest.config ignora .claude/: el worktree guarda una copia del repo y sus
tests corrian duplicados junto a los del arbol actual, inflando el conteo de
la suite y reportando fallos de una rama que no se esta trabajando. El mock
de uuid en CollaboratorUseCases pasa a mockear randomUUID de crypto.

+++ src/modules/catalog/infrastructure/PostgresCatalogRepository.ts
@@ -140,9 +140,12 @@ export class PostgresCatalogRepository implements ICatalogRepository {
      */
     async generateIncrementalId(categoryId: number, prefix?: string): Promise<string> {
         if (prefix) {
+            // El ::int del parámetro no es opcional: sin él Postgres lo toma como
+            // texto y resuelve SUBSTRING(id FROM $1) por su variante de REGEX,
+            // devolviendo el dígito encontrado en vez del sufijo del ID.
             const row = await this.assetRepo
                 .createQueryBuilder('a')
-                .select(`MAX(CAST(SUBSTRING(a.id FROM :len) AS INTEGER))`, 'max')
+                .select(`MAX(CAST(SUBSTRING(a.id FROM :len::int) AS INTEGER))`, 'max')
                 .where('a.id ~ :pattern', { pattern: `^${prefix}[0-9]+$` })
                 .setParameter('len', prefix.length + 1)
                 .getRawOne<{ max: string | null }>();
